Complete rewrite of script
Now takes multiple group names on the command line and reports all members in each.
Output is much more terse.  Code is terser too.  More idiomatic perl.
Seems to work for me, but it is minimally tested, and rather fragile.

For the next release / rewrite, I would like to see:
  - a more flexible and friendly interface
  - -h/--help, -v/--verbose, and --version options
  - output options, not just one size fits all
  - more defensive code (warn and die statements, inline testing).